app/config.py
import os
from dotenv import load_dotenv

load_dotenv('environment.env')

DATABASE_URL = os.getenv("DATABASE_URL")
CELERY_BROKER_URL = os.getenv('CELERY_BROKER_URL')
CELERY_RESULT_BACKEND = os.getenv('CELERY_RESULT_BACKEND')
ACCESS_TOKEN = os.getenv("ACCESS_TOKEN")
REFRESH_TOKEN = os.getenv("REFRESH_TOKEN")
