requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
azure-core==1.30.2
azure-storage-blob==12.20.0
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
cryptography==42.0.8
dask==2024.5.2
databases==0.9.0
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
fsspec==2024.6.0
fuzzywuzzy==0.18.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
isodate==0.6.1
Jinja2==3.1.4
joblib==1.4.2
kombu==5.3.7
Levenshtein==0.25.1
locket==1.0.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
orjson==3.10.3
packaging==24.0
pandas==2.2.2
partd==1.4.2
pluggy==1.5.0
prompt_toolkit==3.0.46
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.3
redis==5.0.4
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.0
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
threadpoolctl==3.5.0
toolz==0.12.1
typer==0.12.3
typing_extensions==4.12.1
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
xgboost==2.0.3
xmltodict==0.13.0
zipp==3.19.2
